+++ src/main/java/niv/burning/api/base/SimpleBurningStorage.java
@@ -118,11 +118,11 @@ public class SimpleBurningStorage extends SingleVariantStorage<FuelVariant> {
         return (oldAmount - newAmount) * resource.getDuration() / newCapacity;
     }
 
-    private static final long clamp(long amount, long min, long max) {
+    private static final int clamp(long amount, long min, long max) {
         if (amount < min)
-            return min;
+            return (int) min;
         else if (amount > max)
-            return max;
+            return (int) max;
         else
             return (int) amount;
     }
